Fix verifica Risposta Aperta API, fix return punteggio e modifiche grafiche generali.
Modificata la visualizzazione degli errori.

Implementazione visualizzazione media dei voti.

Versione 0.7.1-dev

+++ inc/class/Correzione.php
@@ -87,7 +87,7 @@
 			return floor($punteggio_alunno / $punteggio_verifica * 20) / 2;
 		}
 		
-		public function getPunteggioAlunno(int $ID_alunno) : int {
+		public function getPunteggioAlunno(int $ID_alunno) : float {
 			global $mysql;
 			$punteggio = 0;
